refactors project queries
