+++ moped-editor/src/queries/project.js
@@ -1,0 +1,36 @@
+import { gql } from "apollo-boost";
+
+export const PROJECT_NAME = gql`
+  query ProjectName($projectId: Int) {
+    moped_project(where: { project_id: { _eq: $projectId } }) {
+      project_name
+    }
+  }
+`;
+
+export const SUMMARY_QUERY = gql`
+  query ProjectSummary($projectId: Int) {
+    moped_project(where: { project_id: { _eq: $projectId } }) {
+      project_name
+      project_description
+      start_date
+      current_phase
+      current_status
+      capitally_funded
+      eCapris_id
+      fiscal_year
+      project_priority
+    }
+  }
+`;
+
+export const TEAM_QUERY = gql`
+  query TeamSummary($projectId: Int) {
+    moped_proj_personnel(where: { project_id: { _eq: $projectId } }) {
+      first_name
+      last_name
+      role_name
+      notes
+    }
+  }
+`;
