--- moped-editor/src/queries/projects/projectSummaryQuery.js
@@ -1,17 +1,0 @@
-import { gql } from "apollo-boost";
-
-export const SUMMARY_QUERY = gql`
-  query ProjectSummary($projectId: Int) {
-    moped_project(where: { project_id: { _eq: $projectId } }) {
-      project_name
-      project_description
-      start_date
-      current_phase
-      current_status
-      capitally_funded
-      eCapris_id
-      fiscal_year
-      project_priority
-    }
-  }
-`;
