--- moped-editor/src/queries/projects/projectTeamQuery.js
@@ -1,12 +1,0 @@
-import { gql } from "apollo-boost";
-
-export const TEAM_QUERY = gql`
-  query TeamSummary($projectId: Int) {
-    moped_proj_personnel(where: { project_id: { _eq: $projectId } }) {
-      first_name
-      last_name
-      role_name
-      notes
-    }
-  }
-`;
