+++ moped-editor/src/views/projects/projectView/ProjectSummary.js
@@ -1,5 +1,9 @@
 import React from "react";
-import { Box, Grid, CardContent } from "@material-ui/core";
+import { useQuery } from "@apollo/react-hooks";
+import { useParams } from "react-router-dom";
+
+import { Box, Grid, CardContent, CircularProgress } from "@material-ui/core";
+import { SUMMARY_QUERY } from "../../../queries/project";
 
 const formatValue = value => {
   let formattedValue = value.field;
@@ -23,7 +27,16 @@ const formatValue = value => {
   return formattedValue;
 };
 
-const ProjectSummary = details => {
+const ProjectSummary = () => {
+  const { projectId } = useParams();
+
+  const { loading, error, data } = useQuery(SUMMARY_QUERY, {
+    variables: { projectId },
+  });
+
+  if (loading) return <CircularProgress />;
+  if (error) return `Error! ${error.message}`;
+
   const {
     current_status,
     current_phase,
@@ -33,7 +46,7 @@ const ProjectSummary = details => {
     project_priority,
     capitally_funded,
     eCapris_id,
-  } = details.details;
+  } = data.moped_project[0];
 
   const projectDetails = [
     {
